+++ src/Cli.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+
 namespace Senpai;
 
 public static class Cli
@@ -14,7 +15,7 @@ public static class Cli
     /// </summary>
     /// <param name="Args">The args of the application.</param>
     /// <param name="Description">The description of the application.</param>
-    public static int Initialize(string[] Args!!, string Description) => Handler.Invoke(Args, Description, Assembly.GetCallingAssembly());
+    public static int Initialize(string[] Args!!, string? Description) => Handler.Invoke(Args, Description, Assembly.GetCallingAssembly());
 
     /// <summary>
     /// Initiate the command-line interpreter.
@@ -22,5 +23,5 @@ public static class Cli
     /// <param name="Args">The args of the application.</param>
     /// <param name="Description">The description of the application.</param>
     /// <param name="Caller">The assembly that contains the decorated methods.</param>
-    public static int Initialize(string[] Args!!, string Description, Assembly Caller!!) => Handler.Invoke(Args, Description, Caller);
+    public static int Initialize(string[] Args!!, string? Description, Assembly Caller!!) => Handler.Invoke(Args, Description, Caller);
 }