+++ src/Token/Command.cs
@@ -3,50 +3,68 @@ using static Senpai.Exception;
 
 namespace Senpai.Token;
 
-[AttributeUsage(AttributeTargets.Method,
+[AttributeUsage(AttributeTargets.Class,
                 AllowMultiple = false,
                 Inherited = false)]
-public sealed class Command : Symbol
+public sealed class Command : Attribute
 {
     /// <summary>
-    /// Declare a method as a command.
+    /// Declare a class as a command.
     /// </summary>
     /// <param name="Name">The name of the command.</param>
-    /// <param name="Description">The description of the command.</param>
     public Command(string Name,
-                   string? Description                     = default,
-                   [CallerFilePath]   string? __reserved_1 = default,
-                   [CallerMemberName] string? __reserved_2 = default,
-                   [CallerLineNumber] int __reserved_3     = default)
+                   [CallerFilePath] string? _r1 = default,
+                   [CallerMemberName] string? _r2 = default,
+                   [CallerLineNumber] int _r3 = default)
     {
         if (string.IsNullOrWhiteSpace(this.Name = Name))
             throw new ArgumentNullException(nameof(Name));
 
+        if (!string.IsNullOrEmpty(Summary))
+            this.Summary = Summary;
+
         if (!string.IsNullOrEmpty(Description))
             this.Description = Description;
 
-        this.Source = new StackInfo() {
-            File    = __reserved_1,
-            Member  = __reserved_2,
-            Line    = (uint)__reserved_3
+        this.StackTrace = new StackTraceObject()
+        {
+            File    = _r1,
+            Member  = _r2,
+            Line    = (uint)_r3
         };
     }
 
     /// <summary>
-    /// The name of the command.
+    /// A brief description of the command.
     /// </summary>
-    internal override string Name
+    public string Summary
     {
-        get => base.Name;
-        set => base.Name = value;
-    }
+        get;
+        set;
+    } = string.Empty;
 
     /// <summary>
-    /// The description of the command.
+    /// The full description of the command.
     /// </summary>
-    public override string Description
+    public string Description
     {
-        get => base.Description;
-        set => base.Description = value;
+        get;
+        set;
+    } = "No description provided.";
+
+    /// <summary>
+    /// Debugging information about the source of the caller.
+    /// </summary>
+    internal StackTraceObject StackTrace
+    {
+        get;
+        set;
     }
+
+    internal string Name
+    {
+        get;
+        set;
+    } = string.Empty;
+
 }